# chore: Improve performance of table inline editing