--- a/src/table/body-cell/styles.scss
+++ b/src/table/body-cell/styles.scss
@@ -189,6 +189,7 @@ $border-placeholder: awsui.$border-item-width solid transparent;
       }
       & > .body-cell-editor {
         opacity: 0;
+        outline: 0;
       }
       &:hover {
         position: relative;