@gethinwebster 2023-04-19
Keyboard interactions seem to be broken: I can't activate edit on enter (or space), whereas I could previously